backend/tools/mcp_server/pyproject.toml
[project]
name = "mcp_server"
version = "0.1.0"
description = "MCP development server for fullstack-base"
authors = [
    {name = "Finn", email = "finn@claralabs.com"},
]
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
